requirements.txt
streamlit
pandas
numpy
scikit-learn
pillow
requests